add question 5
This one had me stumped as to why it wasn't working so I brought it into eclipse to see if it was a coding issue. The question was worded strangely and it took me a few tries to figure out what it wanted from me. Ended up getting the right answer in the end.